+++ packages/create-app/template-lit-element-ts/vite.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  build: {
+    lib: {
+      entry: 'src/my-element.ts',
+      formats: ['es']
+    },
+    rollupOptions: {
+      external: /^lit-element/
+    }
+  }
+})