+++ packages/vite/src/node/optimizer/esbuildDepPlugin.ts
@@ -3,15 +3,29 @@ import { Plugin } from 'esbuild'
 import { knownAssetTypes } from '../constants'
 import { ResolvedConfig } from '..'
 import chalk from 'chalk'
-import { deepImportRE, isBuiltin } from '../utils'
+import { deepImportRE, isBuiltin, isRunningWithYarnPnp } from '../utils'
 import { tryNodeResolve } from '../plugins/resolve'
+import { PluginContainer } from '../server/pluginContainer'
 
-const externalTypes = ['css', 'vue', 'svelte', ...knownAssetTypes]
+const externalTypes = [
+  'css',
+  // supported pre-processor types
+  'less',
+  'sass',
+  'scss',
+  'style',
+  'stylus',
+  // known SFC types
+  'vue',
+  'svelte',
+  ...knownAssetTypes
+]
 
 export function esbuildDepPlugin(
   qualified: Record<string, string>,
   config: ResolvedConfig,
-  transitiveOptimized: Record<string, true>
+  transitiveOptimized: Record<string, true>,
+  aliasResolver: PluginContainer
 ): Plugin {
   return {
     name: 'vite:dep-pre-bundle',
@@ -32,7 +46,7 @@ export function esbuildDepPlugin(
         }
       )
 
-      build.onResolve({ filter: /^[\w@]/ }, ({ path: id, importer }) => {
+      build.onResolve({ filter: /^[\w@]/ }, async ({ path: id, importer }) => {
         // ensure esbuild uses our resolved entires of optimized deps in all
         // cases
         if (id in qualified) {
@@ -44,6 +58,8 @@ export function esbuildDepPlugin(
           const deepMatch = id.match(deepImportRE)
           const pkgId = deepMatch ? deepMatch[1] || deepMatch[2] : id
           transitiveOptimized[pkgId] = true
+
+          id = (await aliasResolver.resolveId(id))?.id || id
           const resolved = tryNodeResolve(
             id,
             path.dirname(importer),
@@ -86,6 +102,17 @@ export function esbuildDepPlugin(
           }
         }
       )
+
+      // yarn 2 pnp compat
+      if (isRunningWithYarnPnp) {
+        build.onResolve({ filter: /\.yarn.*/ }, (args) => ({
+          path: require.resolve(args.path, { paths: [args.resolveDir] })
+        }))
+        build.onLoad({ filter: /\.yarn.*/ }, async (args) => ({
+          contents: await require('fs').promises.readFile(args.path),
+          loader: 'default'
+        }))
+      }
     }
   }
 }