+++ packages/vite/src/node/server/middlewares/transform.ts
@@ -65,8 +65,7 @@ export function transformMiddleware(
         if (map) {
           return send(req, res, JSON.stringify(map), 'json')
         } else {
-          res.statusCode = 404
-          return res.end()
+          return next()
         }
       }